Drop daily values, add previous hour of rain and hourly wind gust.
The idea of a day is awkward in the context of the station because we don't
necessarily know the time zone, etc... So it's easier to build daily
values after the fact from hourly information.

Also, having previous hour is good because that's probably a more accurate
value than using some arbitrary sample in the current hour.